image_tools/duplicate_finder/requirements.txt
ImageHash==4.3.1
numpy==2.2.0
pillow==11.0.0
PyWavelets==1.8.0
scipy==1.14.1
